fixed some bugs example:"addind empty spaces in command line"
