+++ srcs/parser.c
@@ -6,7 +6,7 @@
 /*   By: gecarval <gecarval@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 08:39:32 by gecarval          #+#    #+#             */
-/*   Updated: 2024/11/08 09:16:51 by gecarval         ###   ########.fr       */
+/*   Updated: 2024/11/12 11:59:53 by gecarval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ void	add_cmd(t_shell *shell, char **args, int is_pipe)
 	if (new == NULL)
 		return ;
 	new->next = NULL;
-	new->cmd = ft_remove_quotes(ft_strdup(args[0]), ft_strlen(args[0]));
+	new->cmd = ft_strdup(args[0]), ft_strlen(args[0]);
 	if (is_pipe == 1)
 		new->type = PIPE;
 	else
